pyproject.toml
[project]
name = "virtualmootcourt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "jsonlines>=4.0.0",
    "load-dotenv>=0.1.0",
    "openai>=1.93.0",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
]
